pyproject.toml
[tool.poetry]
name = "whitelist-bot"
version = "0.1.0"
description = ""
authors = ["qellyka <qellyka@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
py-cord = "^2.6.1"
python-dotenv = "^1.0.1"
mcrcon = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
